Add files via upload
A. Create a javascript arrow function that takes an array of words and transforms it into an array of each word's length and returns it. Display the resulting array on console. (Note:- No test case will contain punctuation. Arrays can be of various lengths) (5 marks)
Example
wordLengths(["hello", "world"]) ➞ [5, 5]
B. Create a javascript function that takes an object and returns the keys and values as separate arrays.(Note: Remember to sort the keys) (5 marks)
Example
keysAndValues({ a: 1, b: 2, c: 3 })➞ [["a", "b", "c"], [1, 2, 3]]
2. Create a node.js application that imports a module which takes first name and last name from the user and copies the full name in a new file “PersonName.txt” as “Your Full Name is <fullname>” following which file is compressed automatically. After copying the content successfully, the page should be redirected to a url “/success” and displays success message on web page and send the status code as 201. (10 marks)
3. Create a node js application which imports a user defined module that reads a file “data.txt” containing some text and returns count of words and characters in that file. Store the result in a separate log file “resultLog.txt” and display on console. In case of file missing or empty, display an error “File not found”. (10 marks